requirements.txt
pandas>=1.3.0
nltk>=3.6.0
googletrans==4.0.0-rc1
matplotlib>=3.3.0
requests>=2.25.0
discord.py>=2.3.0
nest-asyncio>=1.5.0
python-dotenv>=1.0.0
groq-sdk>=0.3.0
